requirements.txt
# Core dependencies
openai>=1.3.0
anthropic>=0.7.0
python-dotenv>=1.0.0

# Document processing
PyPDF2>=3.0.0
python-docx>=1.1.0
markdown>=3.5.0
beautifulsoup4>=4.12.0

# NLP and text processing
tiktoken>=0.5.0
sentence-transformers>=2.2.0
numpy>=1.24.0

# CLI and utilities
rich>=13.7.0
tqdm>=4.66.0

# Output generation
Jinja2>=3.1.2

